Allow mock drivers to be injected
